renamed file process_dicom_Scan_files to ProcessDICOM
