+++ MEDimage/wrangling/__init__.py
@@ -1,4 +1,4 @@
 from . import *
 from .combine_slices import *
 from .DataManager import *
-from .process_dicom_scan_files import *
+from .ProcessDICOM import *
